client and server side page switching illustration
Again we change the clojurescript startup concept.
From now on all sides are initialized with same
core method 'start'. There the page keyword
corresponding to the current URL is determined and
and the appriate page switch is applied on the client
side by hiding all disabled pages and showing the
currently enabled page.

This approach allows to address a page by an URL
(via browser address line) and also from clojurescript
by invoking pages/switch-to-page method. Later on
we can update the URL there but this requires HTML5.

Signed-off-by: Otto Linnemann <linneman@gmx.de>