server and client side page switching
This commit comprises several changes with respect to server and
client side page switching for making it possible to split a complex
website in several subprojects which do not need to be loaded enterily
in the beginning of a session. In this commit we load the pages
index, nav, register and login in the beginning. For demonstration
purposes the profile page is not transfered when the site is initially
visited via the URL index.html. Within the page enable handler, here
enable-profile-page, a page reload is triggered due to the missing
enclosing div element of the profile page.

This commit comprises furthermore:

* added empty status page which is now shown after login
* dom rendering is executed conditionally for many elemetns
  to avoid exceptions when the page has not been transfered yet.
  There might be still errors left in this respect.
* get-element function for restricting dom searches to certain nodes

Signed-off-by: Otto Linnemann <linneman@gmx.de>