implemented sync for profile data
With this changeset we implement exemplarily
the communication between the client and the
server in both directions based on the trans-
mission of json data.

Due to the high update frequency of the profile's
text pane the received html data is first stored
in a cache which is written to the backend database
once a minute. The profile data provides a time
stamp to avoid the rare case of overwriting an
older with a newer data set on the database.

When the profile pane is enabled firstly its
content is updated with the response of an
ajax get request.

Signed-off-by: Otto Linnemann <linneman@gmx.de>