first implementation of search table controller
This version emulates classical html search results
which are presented in a table with fixed number
of rows. If the number of search results exceeds
the number of rows a group index button on the
bottom table allows to skip to the next search
results. If there more search results than columns
within in the search controller which encapsulates
the controll buttons, the search controller is
rerendered.

Signed-off-by: Otto Linnemann <linneman@gmx.de>