first experiments with some clojurescript macros
After having struggeled a long time with macro
integration in clojurescript and the currently
used build system lein-clojurescript I finally
managed to make them work.

Note: At the moment the macro file needs to be
stored twice, once for clojurescript (extention
cljs) the other one for clojure which is used
to precompile a macro (extention clj).

Since clojurescript macros are currently not
integrated within the browser REPL I decided
not to use them as long as they are not
strictly needed. For the given case of the
hash-args function this is obviously not the
case.

Signed-off-by: Otto Linnemann <linneman@gmx.de>