consistent table controller row height
In case not all rows are rendered by the table controller
html adjusts the row height in order to fit to the table
height which is not intended here. The only way to surround
this seems to be always to render the maximum amount of
table rows and set them to a constant height. Both changes
are introduced with the present commit.

Signed-off-by: Otto Linnemann <linneman@gmx.de>