implemented multilanguage support
This commit introduces support for several languages. One of the big
advantages of the given design is that supporting a new language is
in principle relatively simple due to the strict separation of
code, style and text content. Nevertheless this is a big change with
respect to the number of functions/files which had to been patched.

Multilanguage support works in following way: The first resource
locator after the domain name forms the language code e.g. "de" for
German, "en" for English, etc. The remaining parts of HTTP/AJAX
requests remain unchanged. The only language depended information
resides in the directory /public/templates which is used to deliver
static html pages so only the compojure-routes (server) to html
pages require language addressing. Instead of using matching
rules for all the other routes it seems to be more straigt forward
to rewrite the url for all the request which are language independent
and to store the requested language within the compojure request
parameter :lang. The advantage of this approach is that the posted
parameters remain unchanged. Furhtermore we use this so called
rewrite handler to redirect to URL's to a default language in case
the language code is not given.

Example (when German is default language):

(HTTP): http://<domain>/index.html
        -> rerirect to -> http://<domain>/de/index.html

(AJAX): http://<domain>/en/login
        -> (:lang=en) -> http://<domain>/login

(AJAX): http://<domain>/login
        -> (:lang=de) -> http://<domain>/login

Signed-off-by: Otto Linnemann <linneman@gmx.de>