security bug fix within authentication handler
The white list filtering within the authentication handler
was not working correctly when the initial method name
of some white-list enteties matched the initial portion
of the request uri. In example the requst

/user-matches
was not filtered out because the request for the getting
the users name was defined within the white-list as /user.
By changing the regular expression this was solved.

Furthermore the handler sequence was reordered to allow
a language change at any time. It is now invoked before
the session and authentication handlers to ensure that
the cookie data is correctly addressed with language
prefix. ( I do not know how to get rid of the language
prefix for the cookies but at least I can ensure that
the rewrite handler is invoked before the rest. )

Signed-off-by: Otto Linnemann <linneman@gmx.de>