implemented profile integrity check
Before search results are displayed and the users level is increased
within table user all required profile data have to be specified. If
this is not the case and the user clicks the search pane, instead of
the search results a notification text is displayed given directions
to complete the profile data.

Handling the initial sequence of entering profile data which is put
into the servers cache and then switching immediately to the search
pane is somehow tricky because the cached profile data has to be
stored into the database and check for integrity before the matching
function is actually triggered. Within the present implementation
introduced with this commit we rely on a delayed post request to ensure
that the cached profile data is actually processed. While this approach
is not verly elegant it seems to reliably work. If in some rare cases
the profile cache shall not processed fast enough the user has just
to switch the tabe pandes (profile, search) again.

Signed-off-by: Otto Linnemann <linneman@gmx.de>