corrected bug when favorite member deleted account
The LEFT JOIN SQL request for comibining three tables delivered
NULL enteties in this case. For this reason LEFT JOIN is
replaced by regular JOIN.

In case a user deletes his profile data which is part
of the favorite list of another user the delete
user profile is not displayed on the users favorite
list anymore.

Signed-off-by: Otto Linnemann <linneman@gmx.de>