flexible number of profile questions
Due to some limitations of SQL the number of
predefined questions in the profile questionaire
was exactly 10. With the present commit we
try to arrange this number more flexible by
defining an unique constant of supported
questions on the server side (file setup).

This allows to reduce the number of required
questions in a simple way while the matching
variance is still correctly calculated.

Be aware that the client side has still a hard
coded value of maximum 10 questions in several
location. This will be changed if required later
on.

Signed-off-by: Otto Linnemann <linneman@gmx.de>