implemented language change methods on client
Switching between languages just by reloading
the current side with another language prefix
is dangerous due to language dependend cookies.
When the user logs out and reloads the page
with a previously used language prefix where
he did not explicitly perform a logout operation
the side hangs because the server is in logged
out state while client still assumes of being
correctly logged in due to outdated cookies.

In order to allow to change language in more
realiably we introduce a client side clojure-
script method which clears all cookies before
reloading the page with a new language prefix.

This approach fails of course if the user
changes languages via the address line and
logs out in between. The only way to get
the page working again would be to clear
the cookies manually or to login via the
language where last logout took place.
Hopefully this case does not occur to often.

Signed-off-by: Otto Linnemann <linneman@gmx.de>