more robust language switch
This commit introduces an additional fall back
mechanism to avoid that client and server run
out of sync which could still happen in case
of language switch (see previous commits).

When the client is not authenticated an appropriate
HTTP error code is now returned for AJAS requests
and the authentication cookie is deleted. This
is achieved by setting cookie arguments within
the ring response handler directly. The ring
framework was upgraded to version 1.1.0 because the
newer version  provides a convenient cookie response
wrapper in order to do this.

On the client side the afore mentioned http error
code 403 is now handled before the ajax callback
function is called. The ajax function "send-request"
was therefore shifted from the utils namespace to
a the the namespace ajax. When error 403 is catched
the cookie used to signal authentication is cleared
and the index page is invoked in order to allow
users to correctly log in.

Signed-off-by: Otto Linnemann <linneman@gmx.de>