improvements of runtime system for development
Changes which allow to use the same clojure
version (from this project) for server and
client side repl based development. A repl
can now started without the webserver and
additional tasks with the command "lein repl".
The clojure command (cljs-repl) opens an
clojurescript repl which connects to the
client on the browser. The command 'lein run'
works the same way as before. It activates all
services including a swank server for allowing
to connect a slime client e.g. in emacs.

Signed-off-by: Otto Linnemann <linneman@gmx.de>