complex db requests for message info
Up to this commit the message pane displayed distance
and matching score for each user leading to the need
for somehow complex sql requests. While this is possible
these request cause lots of redundant code and
ultimately the information might not be strictly needed.
So we postphone this implementation and go remove it
for now.

Signed-off-by: Otto Linnemann <linneman@gmx.de>