fundamental changes for fixed css layout
Using a fixed width layout has several advantages. First and formost
the site is displayed correctly even on mobile devices. But css
formats with a fixed width simply the design of a webapplication, too.
Only for more informative sites such as wikipedia a floating layout
does really make sense in this respect. For that reason we change
the layout to the overall width of 960 pixels which seems to be
more or less standard. The logout and impress button are shifted to
the status line which gives only three major states for messages
(status information), profile data and search.

Signed-off-by: Otto Linnemann <linneman@gmx.de>