do not clean registered state after lang change
Cookie and session data handling is still subject
for further improvements. Unfortunately cookies are
attached to the url prefix. For this reason we get
new session data when the language is changed which
is not intentended but I currently do not know how
to change this.

With this commit we at least avoid to display the
registration button after language change when the
user was registered before. Maybe we can avoid
deregistering after language change at all by some
server tweaks.

Signed-off-by: Otto Linnemann <linneman@gmx.de>