several fixes and improvements in authentication
Commit 59ed4b33a59f18f68c39c4163b4382ce9f61891c
(fix confirmation link generation bug) addressed
allready the problem of plus signs generated
for url argument keys which are replaced by blans
and therefore cause authentication to always
fail if it happens. For this reason the function
url-encode-key was patched by a new version.
Unfortunately I forgot to replace the actual
invocation by the patched version so the bug
was actually not fixed. This should hopefully
be clarified now.

Futhermore cookie generation is added to the
wrap-authentification handler which allows to
more consistently handle registered and
authenticated operation modes. Especially when
switching languages we had to clean all
cookies because otherwise mismatches between
the session state on the server and the cookie
state appeared very often. Now the more reliable
session state is used to always update the
cookies. The updating mechanism in the
POST handlers for login, logout, registering
and confirmation are nevertheless required since
the client does generally not reload the page
after these methods have been invoked via
an ajax request. So the cookie handling within
auth.clj seems to be redundant but is actually
required in this way until the client side
is not updated.

Signed-off-by: Otto Linnemann <linneman@gmx.de>