fix sync issue for lon/lat calculation
When the running the database middlewhere process
which calculates latitude and longitude for a given
users zip asynchronously the calculation could still
not be finished when the user already triggers a
search operation which in turn runs on wrong data.
Thus we have to wait until the database operation
have been finished whithin function check profile.
This fix is introduced with the present commit.

Furthermore a minor fix is introduced to avoid the
concatenation of the https port when the default
port 443 is used.

Signed-off-by: Otto Linnemann <linneman@gmx.de>