add last_seek column in profile table
which is updated whenever a new matching list
is retrieved from database. This allows later
on to automtically inform users about not
yet discovered potential matches via an email
notification.

Furhtermore a bug which caused the modified
timestamp in the users profile table always
to be updated is fixed.

Pay attention to the fact that the client side
currently updates the users profile ways too
often causing the profile modification date
beeing altered whenever the user selects
the being found tab. This is a known problem
and needs to be addressed on the client. At
the given moment the profile modification date
is not used so this is a minor problem.

Signed-off-by: Otto Linnemann <linneman@gmx.de>