email notification about new matches and new messages
Keeping users informed about what is going on in the project is
crucial when the side is going online first. Otherwise we have
to take into account that most users will create an account for
the side, search for matches and after having nothing interesting
found never come back again. So especially when going online for
the first time we need to inform the few users we can hopefully
convince to create an account about potential new matches via
email providing a link where they directly get these information
in the web interface.

This is realized by a background process started around midnight.
It determines all users who have received new messages since the
last visit and all users where new matching profiles are available
the user has not retrieved yet. Especially the last point is a
little bit tricky to implement. Maybe there are still some bugs.

Since all messaging has to be done in the users prefered language
we have to store the last address language in the users profiles
table of the database, too. This is done by another http handler
in core.clj.

Signed-off-by: Otto Linnemann <linneman@gmx.de>