notifier pane when no matches available
When the side is launched there are no members registered
at all. We want to communicate this in an open way so there
is a special information text display explaining the
situation, apply for telling others, etc. These information
pane includes also the support button from the start page
where unfortunately some code is duplicated. Since this is
only temporarily needed the changes in this commit can be
reverted later on.

Signed-off-by: Otto Linnemann <linneman@gmx.de>