fix for scrolling in editor for iOS devices
added the css property "-webkit-overflow-scrolling: touch" to
somehow improve the missing scrolling capability in googles
editor when embedded in another iframe e.g. tabpane or
dialog box. Unfortunately there are more bugs with iOS. In
example the tabpane pictures within the profile tabs are not
drawn with correct z-index on ipad anymore. Nevertheless in
order to somehow allow usage on iPad with decide to use the
aforementioned properties.

A workaround for graphics would be to programatically draw
them by some javascript. We will see, maybe Apple will fix
this some day.

Signed-off-by: Otto Linnemann <linneman@gmx.de>