special treatment for 4th profile question
The 4th question: 'Are you willing to move?' gets a special treatment.
Unfortunately the expression to be used is obviously too complex to
to be calculated inside MySQL so we postprocess the database result
and do not include the calculation for question 4 inside the database.
Refer to function match-variance-for for more detailed information
about how we treat this question.

Signed-off-by: Otto Linnemann <linneman@gmx.de>