correct misleading description in message menu
Contact requests without reply can be empty  even when an user has already send
out some messages. It is exclusivley introduced to ensure that a user can track
all messages which  belong to him. Since messages are  generally grouped within
communication streams the ones which have  not been replied yet cannot assigned
to such a stream and are therfore  only temporarily grouped under the third tab
pane. Maybe we need  a better description of how this is  considered to work on
the page. Be careful not to overcharge users with a too technical description.

Signed-off-by: Otto Linnemann <linneman@gmx.de>