improved rendering of user details dialog
The 8th answer "are you very tall?" needs to be specified
by the size range instead of the general statement
"fully applies, partially applies, etc". The user details
dialog should be generally improved to include the text
from the pop up menues of the questionnaire but requires
much more efford for the synthesis of the correct answer
phrases. Much probably we need all these messages twice.
For this reason we just keep it simple for the moment.

Signed-off-by: Otto Linnemann <linneman@gmx.de>